Don't re-issue proxy cookie when refreshing the session (#3)
Re-issuing the proxy cookie could change the expiration if the upstream
uses `Max-Age` rather than a fixed expiration time. We could re-compute
`Max-Age` by storing the cookie issuance time in the encrypted proxy
cookie value, but it's simpler to just not touch the proxy cookie except
when the upstream issues or changes it.

Also fixes an issue where we weren't sending a new challenge in response
to an invalid challenge in the refresh endpoint.